Check if autoload.php exists
Hi,

when installing this plugin/package via composer the dependencies will be located outside the plugin folder. A simple check if the `autoload.php` exists would make it work in such environments as well (Roots Bedrock in my case) because `wamania/php-stemmer` is already installed in the global vendor folder.